refactor: Extend test shouldReturnIdWithSizeMatchingRequest
