+++ src/test/java/no/sikt/nva/pubchannels/handler/search/BaseSearchSerialPublicationByQueryHandlerTest.java
@@ -3,13 +3,11 @@ package no.sikt.nva.pubchannels.handler.search;
 import static java.util.Objects.nonNull;
 import static no.sikt.nva.pubchannels.TestConstants.DEFAULT_OFFSET_INT;
 import static no.sikt.nva.pubchannels.TestConstants.DEFAULT_SIZE_INT;
-import static no.sikt.nva.pubchannels.TestConstants.ISSN_QUERY_PARAM;
 import static no.sikt.nva.pubchannels.TestConstants.NAME_QUERY_PARAM;
 import static no.sikt.nva.pubchannels.handler.TestUtils.constructPublicationChannelUri;
 import static no.sikt.nva.pubchannels.handler.TestUtils.constructRequest;
 import static no.unit.nva.testutils.RandomDataGenerator.objectMapper;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsStringIgnoringCase;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -25,8 +23,6 @@ import no.unit.nva.commons.pagination.PaginatedSearchResult;
 import nva.commons.apigateway.GatewayResponse;
 import nva.commons.apigateway.exceptions.UnprocessableContentException;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 public abstract class BaseSearchSerialPublicationByQueryHandlerTest extends SearchByQueryHandlerTest {
 
@@ -69,27 +65,6 @@ public abstract class BaseSearchSerialPublicationByQueryHandlerTest extends Sear
         var result = objectMapper.readValue(response.getBody(), TYPE_REF).getHits().getFirst();
 
         assertThat(result.year(), is(equalTo(year)));
-    }
-
-    @ParameterizedTest(name = "Should accept size \"{0}\"")
-    @ValueSource(ints = {1, 10, 20})
-    void shouldReturnIdWithSizeMatchingRequest(int size) throws IOException {
-        var testChannel = new TestChannel(year, pid, type).withPrintIssn(issn);
-        mockChannelRegistryResponse(year,
-                                    ISSN_QUERY_PARAM,
-                                    testChannel.getPrintIssnValue(),
-                                    List.of(testChannel.asChannelRegistrySerialPublicationBody()),
-                                    size,
-                                    DEFAULT_OFFSET_INT);
-
-        var input = constructRequest(Map.of("year", year, "query", issn, "size", String.valueOf(size)),
-                                     MediaType.ANY_TYPE);
-        handlerUnderTest.handleRequest(input, output, context);
-
-        var response = GatewayResponse.fromOutputStream(output, PaginatedSearchResult.class);
-        var actualId = objectMapper.readValue(response.getBody(), TYPE_REF).getId();
-
-        assertThat(actualId.toString(), containsStringIgnoringCase("size=" + size));
     }
 }
